updated pXsec to do analysis.
about 80 lines of real code added to ProtonXsec to do the analysis.

added to the BeamSelector class 2 functions. getMCInitialKE and
getDataInitialKE. Right now the data version returns whatever you feed
it -> needs to become related to WC track momentum + energy loss in the
front face of the TPC. The MC version mimicks the MC analysis.

Did a quick test on the distributions generated by ProtonAnalyzerMC and
ProtonXsec and they look the same.

Did not yet add any corrections read in from the MC file. Eventually
would be nice to clean up how the root files are written out to make
things cleaner.